requirements.txt
python-telegram-bot[webhooks]>=20.0
openai>=0.27.0
sqlalchemy>=2.0
psycopg2-binary>=2.9


